feat: make autoRefresh optional
