+++ src/module.ts
@@ -8,40 +8,7 @@ import {
 } from '@nuxt/kit'
 import { joinURL } from 'ufo'
 import * as DirectusSDK from '@directus/sdk'
-
-export interface ModuleOptions {
-  /**
-   * Directus API URL, customizable at runtime via NUXT_PUBLIC_DIRECTUS_URL environment variable.
-   * @default ''
-   * @type string
-   */
-  url: string;
-  /**
-   * Directus static token.
-   * @default ''
-   * @type string
-   */
-  staticToken?: string;
-  /**
-   * Enable Directus Devtools
-   * @default false
-   * @type boolean
-   * @see https://docs.directus.io/guides/developer-tools.html
-   */
-  devtools?: boolean;
-  /**
-   * Token Cookie Name
-   * @default 'directus_access_token'
-   * @type string
-   */
-  tokenCookieName?: string;
-  /**
-   * Refresh Token Cookie Name
-   * @default 'directus_refresh_token'
-   * @type string
-   */
-  refreshTokenCookieName?: string;
-}
+import { ModuleOptions } from './runtime/types'
 
 export default defineNuxtModule<ModuleOptions>({
   meta: {
@@ -56,7 +23,8 @@ export default defineNuxtModule<ModuleOptions>({
     staticToken: '',
     devtools: false,
     tokenCookieName: 'directus_access_token',
-    refreshTokenCookieName: 'directus_refresh_token'
+    refreshTokenCookieName: 'directus_refresh_token',
+    autoRefresh: true,
   },
   setup (options, nuxt) {
     const { resolve } = createResolver(import.meta.url)
@@ -68,7 +36,8 @@ export default defineNuxtModule<ModuleOptions>({
         staticToken: options.staticToken,
         devtools: options.devtools,
         tokenCookieName: options.tokenCookieName,
-        refreshTokenCookieName: options.refreshTokenCookieName
+        refreshTokenCookieName: options.refreshTokenCookieName,
+        autoRefresh: options.autoRefresh
       }
     )
 
@@ -85,7 +54,10 @@ export default defineNuxtModule<ModuleOptions>({
     nuxt.options.build.transpile.push(runtimeDir)
 
     addPlugin(resolve(runtimeDir, './plugins/directus'), { append: true })
-    addPlugin(resolve(runtimeDir, './plugins/autoRefresh'), { append: true })
+
+    if (nuxt.options.runtimeConfig.public.directus.autoRefresh) {
+      addPlugin(resolve(runtimeDir, './plugins/autoRefresh'), { append: true })
+    }
 
     addImportsDir(resolve(runtimeDir, 'composables'))
